ステージ修正

+++ GameTemplate/Game/GameTime/GameTime.h
@@ -29,7 +29,7 @@ namespace smEngine {
 			::QueryPerformanceFrequency((LARGE_INTEGER*)&m_fleam);
 			::QueryPerformanceCounter((LARGE_INTEGER*)&m_stop);
 			m_elapsed = double(m_stop - m_start) / m_fleam;
-			SetFrameDeltaTime(m_elapsed);
+			SetFrameDeltaTime((float)m_elapsed);
 			//::QueryPerformanceCounter((LARGE_INTEGER*)&m_start);
 		}
 		float GetFrameTime()
